#165: Make YdbValidator implementation customizable via YdbRepository settings
This opens the door for error handling customization that can work without changing the core error handling logic in repository-ydb-v2.
