+++ repository-ydb-v2/src/test/java/tech/ydb/yoj/repository/ydb/YdbSpliteratorTest.java
@@ -8,6 +8,7 @@ import tech.ydb.core.Status;
 import tech.ydb.core.StatusCode;
 import tech.ydb.yoj.repository.db.exception.DeadlineExceededException;
 import tech.ydb.yoj.repository.db.exception.QueryInterruptedException;
+import tech.ydb.yoj.repository.ydb.client.YdbValidator;
 
 import java.time.Duration;
 import java.util.ArrayList;
@@ -73,7 +74,8 @@ public class YdbSpliteratorTest {
     @Test
     @SneakyThrows
     public void endStreamWhenSupplerOfferValue() {
-        YdbSpliterator<Integer> spliterator = new YdbSpliterator<>("stream", false, Duration.ofMillis(500));
+        Duration timeout = Duration.ofMillis(500);
+        YdbSpliterator<Integer> spliterator = new YdbSpliterator<>("stream", YdbValidator.DEFAULT, false, timeout);
 
         spliterator.onNext(1);
 
@@ -198,7 +200,7 @@ public class YdbSpliteratorTest {
         }
 
         public static ReadTableMock start(Duration timeout) {
-            YdbSpliterator<Integer> spliterator = new YdbSpliterator<>("stream", false, timeout);
+            YdbSpliterator<Integer> spliterator = new YdbSpliterator<>("stream", YdbValidator.DEFAULT, false, timeout);
 
             ReadTableMock mock = new ReadTableMock(spliterator);
             mock.run();
